backend/pyproject.toml
[project]
name = "baby-leaps-ai"
version = "0.1.0"
description = "API especialista em saltos de desenvolvimento de bebês com Gemini AI"
authors = [
    { name = "Vitor Oliveira", email = "vitorroliveiraa@outlook.com.br" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "google-genai (>=1.15.0,<2.0.0)"
]

[tool.poetry]
packages = [{ include = "baby_leaps_ai", from = "src" }]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
